Fleshed out PDL::Parallel::threads::SIMD
This includes changing the invocation method from launch_simd to
parallelize, and chaning the order of the arguments for that
function. I also changed the way in which the thread id is
accessed, and I changed the name of the barrier synchronization
function to parallelize_sync.

Also, it includes a complete write of the documentation for
PDL::Parallel::threads::SIMD. I believe this module is now ready
to go, except that it needs tests, too.