Made the module installable on systems without ithreads (hopefully)
After getting erroneous feedback from the cpan testers, I decided to
make this module installable even on perl's that don't have ithreads.
The reason is that PDL::Parallel::threads does not actually need
ithreads support to work, so the user should be allowed to install
the module even if they don't have ithreads.

Now, PDL::Parallel::threads::SIMD will fail on systems that don't
have ithreads, but that's not the main module here. :-)