Compiles under PDL newer than 2.057
A commit en route to PDL v2.058 changed the name of the pdl struct
field pointing to the parent of a transformation piddle. If that
field is null, the piddle is not a slice, and so I checked it
internally. After the name change, the module no longer compiled.
This commit fixes that, though I would prefer to make it work for
both older and newer versions of PDL.