Fix test 30, the one that actually tests sharing across threads.
I just happened to notice a line in the commit log indicating that
a number of PDL functions that used to return scalars now return
1-element piddles. That is almost never a probelm, except when you
try to put the result of all() into a shared array! Naturally, that's
excactly how I test things. Bah. Call sclr on the returned result.